Major cleanup of Labview code.  Ripped out the "Kalman demodulator" code and the ASIO sound card support.  Lots of general cleanup.
Changed the carrier amplitude representations to be consistent keep a "coupling matrix" as the last two dimensions, and added a first dimension for the "Card", to support multiple input cards.  Currently this is set up to just use two cards, the current feedback and on sensor card.  If there are multiple sensor cards you can pick one to process input from.

"Carrier levels" now holds the full coupling matrix for current sense, as card 0, while the sensor card is card 1.  This is to support cancellation of off-axis signals on the source.